Refactor so Agent generates Episodes and only uses for recording

+++ agent/reward_state_action.py
@@ -3,9 +3,10 @@ from typing import Optional
 
 import environment
 
+
 @dataclass
 class RewardStateAction:
-    reward: Optional[int]
+    reward: Optional[float]
     state: Optional[environment.State]
     action: Optional[environment.Action]
